+++ NEWS.md
@@ -5,6 +5,9 @@ CHANGES
 * `demean()` (and `degroup()`) now also work for nested designs, if argument
   `nested = TRUE` and  `by` specifies more than one variable.
 
+* `data_modify()` now also accepts the short-cut `1:n()` and similar, for example
+  to create an index for data groups (in grouped data frames).
+
 # datawizard 0.12.2
 
 * Remove `htmltools` from `Suggests` in an attempt of fixing an error in CRAN